// src/lib/types.ts
export type Education = {
	institution: string
	duration: string
	concentration: string
}

export type Job = {
	title: string
	company: string
	duration: string
	description: string
}

export type Project = {
	link: string
	title: string,
	description: string
}

export type Skill = {
	name: string,
	mastery: number
}